ruby-kafka can't handle different consumer topic subscription in one consumer group ISSUE-525/764
Ruby kafka does not support having consumers subscribed to different topics when they
belong to the same consumer group. One of the reasons this happens is because the default
round robin partition assignment strategy makes a hard assumption that all topic subscriptions
within a consumer group are equivalent.

We're providing our own 'custom' round robin strategy at the moment we initialize the consumer
to circumvent this issue. This strategy is heavily inspired by the original kafka client
[RoundRobinAssignor](https://github.com/apache/kafka/blob/trunk/clients/src/main/java/org/apache/kafka/clients/consumer/RoundRobinAssignor.java)

Finally, we're decided to do this change here in phobos, as opposed to put a PR directly to
ruby-kafka, because we (Flipp) have more control over this project.